LSH: Remove fallback mode
If mask_current=True, then will always attend to the current pos if nothing else is possible.
If mask_current=False, and mask_different_hashes=True, then will attend to different
hashes if nothing else is possible.
Otherwise, will attend to entire key window (also to positions which correspond to masking or the future).
This is necessary s.t. we do not get NaN attention weights.